+++ src/scmdata/timeseries.py
@@ -10,6 +10,8 @@ import functools
 from typing import Any, Callable, List, Union
 
 import numpy as np
+import openscm_units.unit_registry as ur
+import pint
 import xarray as xr
 from xarray.core.ops import inject_binary_ops
 
@@ -205,12 +207,22 @@ class TimeSeries:
         def func(self, other):
             other_data = getattr(other, "_data", other)
 
-            ts = (
-                f(self._data, other_data)
-                if not reflexive
-                else f(other_data, self._data)
-            )
+            if isinstance(other, pint.Quantity):
+                try:
+                    self_data = self._data * ur(self.meta["unit"])
+                except KeyError:
+                    # let Pint assume dimensionless and raise an error as
+                    # necessary
+                    self_data = self._data
+            else:
+                self_data = self._data
+
+            ts = f(self_data, other_data) if not reflexive else f(other_data, self_data)
             ts.attrs = self._data.attrs
+            if isinstance(other, pint.Quantity):
+                ts.attrs["unit"] = str(ts.data.units)
+                ts.data = ts.values
+
             return TimeSeries(ts)
 
         return func